BREAKING CHANGE:Streamline to align better with jwk api.
BREAKING CHANGE!

A few of the options go away in favor of leaning on the jws/jwk library
and jwskeychain to provide more of that verification.

Removed:
- RequirePolicies
- Require
- TrustRootCAs
- Verifier
- SignWith
- SignWithRaw

Added:
- Signer
- NewSigner
- SignWithX509Chain
- SignWithKey
- Decoder
- NewDecoder
- WithKeyProvider
- WithKeySet
- WithKeyUsed
- WithVerifyAuto

This was done to incorporate the flexability available from the jws/jwk
libraries but limit some of the things that could be configured that
would break this library.